+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
-## [4.4.2] - 2026-08-27
+## [4.4.2] - 2026-09-01
 
 ### Fixed
 
 - Refund modal blocked on loading
+- Fixed 500 error on module configuration page
+- Keep GitHub update check throttle when the API request fails
 
 ## [4.4.1] - 2026-08-10